% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmx.R
\name{convert_zmx}
\alias{convert_zmx}
\title{Convert ZMX to OMX}
\usage{
convert_zmx(zmx, omx = NULL)
}
\arguments{
\item{zmx}{Path to \code{.zmx} file.}

\item{omx}{Optional path to new omx file. If \code{NULL}, will save in same
directory with same name as \code{zmx}.}
}
\description{
Convert ZMX to OMX
}
\author{
Ben Stabler
}

