% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_omx.R
\name{create_omx}
\alias{create_omx}
\title{Create an OMX file}
\usage{
create_omx(file, numrows, numcols, level = 1)
}
\arguments{
\item{file}{Full path name of the OMX file to create.}

\item{numrows}{Number of rows that all matrices in the file will have.}

\item{numcols}{Number of columns that all matrices in the file will have.
Defaults to \code{numrows}.}

\item{level}{Compression level. Default \code{1}.}
}
\value{
An active connection to \code{file}
}
\description{
This function creates an OMX file, establishes the shape attribute (number
of rows and columns) and version attribute, and creates the data and lookup
groups.
}

