% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_matrix.R
\name{gather_matrix}
\alias{gather_matrix}
\title{Gather the destinations of a Origin-Destination matrix}
\usage{
gather_matrix(matrix, value_name = NULL)
}
\arguments{
\item{matrix}{An R matrix with row and column names or indices}

\item{value_name}{A character string identifying the name of the value column in
the output data_frame}
}
\value{
A \code{data_frame} with row and column ids and matrix values.
}
\description{
Gather the destinations of a Origin-Destination matrix
}
