% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{get_omx_attr}
\alias{get_omx_attr}
\title{Read OMX attributes}
\usage{
get_omx_attr(file)
}
\arguments{
\item{file}{full path name of the OMX file being read}
}
\value{
A list containing \code{SHAPE} and \code{VERSION} attributes.
}
\description{
This function reads the SHAPE and VERSION attributes of an OMX file. This is
called by several other functions.
}
\examples{
omxfile <- omxr_example("skims.omx")
get_omx_attr(omxfile)

}
