% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omxr.R
\docType{package}
\name{omxr}
\alias{omxr}
\title{\code{omxr}: A package for using OMX files in R.}
\description{
The Open Matrix (OMX) file type is an open specification for 
sharing data from transportation models. The specification is built on  
HDF5 \url{https://www.hdfgroup.org/};  
APIs to read and write are available for 'Cube', 'Emme', 'Python', 'Java',  
and 'R'.
}
