% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omxr.R
\name{omxr_example}
\alias{omxr_example}
\title{Get path to example file}
\usage{
omxr_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
omxr comes bundled with an example omx and zmx file in its `inst/extdata`
directory. This function makes it easy to access.
}
\examples{
omxr_example()
omxr_example("test.omx")
}
