% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{read_lookup}
\alias{read_lookup}
\title{Function to read an OMX lookup}
\usage{
read_lookup(file, name)
}
\arguments{
\item{file}{Path name of the OMX file where the lookup resides.}

\item{name}{Name of the lookup in the OMX file}
}
\value{
A list of two elements:
 \describe{
   \item{\code{lookup}}{The lookup vector.}
   \item{\code{lookup_dim}}{String, whether the lookup refers to rows or
     columns. }
 }
}
\description{
This function reads a lookup and its attributes.
}
