% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmx.R
\name{read_zmx}
\alias{read_zmx}
\title{Read a PB ZMX file}
\usage{
read_zmx(fileName)
}
\arguments{
\item{fileName}{Path to a ZMX file.}
}
\value{
An R matrix object.
}
\description{
Read a PB ZMX file
}
\details{
The ZMX was a PB-designed precursor to OMX. This function is
  provided for backwards compatibility only.
}
\author{
Ben Stabler, Brian Gregor
}

