% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{write_matrix_attr}
\alias{write_matrix_attr}
\title{Function to write matrix attribute}
\usage{
write_matrix_attr(file, name, attr_name, value)
}
\arguments{
\item{file}{Full path name of the OMX file to store the matrix in. If this is
a new matrix file, see \link{create_omx}.}

\item{name}{Name of the matrix in the OMX object.}

\item{attr_name}{Name of the attribute.}

\item{value}{Attribute value}
}
\description{
Function to write matrix attribute
}
