% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmx.R
\name{write_zmx}
\alias{write_zmx}
\title{Write a PB ZMX file}
\usage{
write_zmx(Matrix, zmx, sevenz = NULL)
}
\arguments{
\item{Matrix}{An R matrix}

\item{zmx}{Path to a ZMX file.}

\item{sevenz}{Path to 7zip executable. Defaults to \code{SEVENZIP_BIN} in
\code{Sys.getenv()}}
}
\value{
An R matrix object.
}
\description{
Write a PB ZMX file
}
\details{
The ZMX was a PB-designed precursor to OMX. This function is
  provided for backwards compatibility only.
}
\author{
Ben Stabler, Brian Gregor
}
