% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_matrix.R
\name{long_matrix}
\alias{long_matrix}
\title{Transform a matrix into long format}
\usage{
long_matrix(matrix)
}
\arguments{
\item{matrix}{An R matrix with row and column names or indices}
}
\value{
A \code{data_frame} with row and column ids and matrix values.
}
\description{
Transform a matrix into long format
}

